🚧 Enhanced entities + created SQL-statements for vaccine type + vaccine schedule.
BEFORE creating NotificationEntity.

+++ src/main/java/lu/pokevax/technical/dialect/EntityContract.java
@@ -2,5 +2,10 @@ package lu.pokevax.technical.dialect;
 
 public interface EntityContract {
 
-    Long getId();
+    /**
+     * Why not {@link Long} ? was convenient with SQL-lite to use AUTOINCREMENT feature.
+     *
+     * @return db identifier of the entity.
+     */
+    Integer getId();
 }
